# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/app/config/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "更新app的配置文件",
                "parameters": [
                    {
                        "description": "更新app的配置文件",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppConfigUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/config/{app_name}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "获取app配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_name",
                        "name": "app_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppConfigRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app创建",
                "parameters": [
                    {
                        "description": "app创建",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app删除",
                "parameters": [
                    {
                        "description": "app删除",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/detail/{app_name}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "获取app详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_name",
                        "name": "app_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppDetailRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/env": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app更新env",
                "parameters": [
                    {
                        "description": "app更新env",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppEnvReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app信息列表",
                "parameters": [
                    {
                        "description": "app信息列表",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppListRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/open": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app开启或关闭集群外访问",
                "parameters": [
                    {
                        "description": "app开启或关闭集群外访问",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppOpenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/replicas": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app副本数伸缩",
                "parameters": [
                    {
                        "description": "app副本数伸缩",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppReplicasReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/restart": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app重启",
                "parameters": [
                    {
                        "description": "app重启",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppRestartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/restore": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app根据快照恢复数据",
                "parameters": [
                    {
                        "description": "app根据快照恢复数据",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppRestoreReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/sc": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "获取支持的存储类型",
                "responses": {
                    "200": {
                        "description": "获取支持的存储类型",
                        "schema": {
                            "$ref": "#/definitions/model.SCListRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/scale": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app扩缩容",
                "parameters": [
                    {
                        "description": "app扩缩容,主要是内存和cpu的调整",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppScaleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
...
        "/app/start": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app启动",
                "parameters": [
                    {
                        "description": "app启动",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppRestartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/app/stop": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app停止",
                "parameters": [
                    {
                        "description": "app停止",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppRestartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AppConfigBase": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "model.AppConfigRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.AppConfigBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.AppConfigUpdateReq": {
            "type": "object",
            "required": [
                "data",
                "name"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "app名称",
                    "type": "string"
                }
            }
        },
        "model.AppCreateReq": {
            "type": "object",
            "required": [
                "app_little_category",
                "cpu",
                "mem",
                "show_name",
                "tag"
            ],
            "properties": {
                "app_little_category": {
                    "description": "app小类",
                    "type": "string"
                },
                "cpu": {
                    "description": "cpu核心数，例如1，2，3，4",
                    "type": "string"
                },
                "env": {
                    "description": "环境变量，格式key:value,例如[{\"ROOT_PASSWORD:2222\"},{\"AGE:4444\"}]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mem": {
                    "description": "内存大小，例如1，2，3，4",
                    "type": "string"
                },
                "sc": {
                    "description": "储存类型",
                    "type": "string"
                },
                "show_name": {
                    "description": "名称",
                    "type": "string"
                },
                "tag": {
                    "description": "版本号",
                    "type": "string"
                },
                "volume_size": {
                    "description": "磁盘大小",
                    "type": "string"
                }
            }
        },
        "model.AppDeleteReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.AppDetailRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.AppDetailBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.AppEnvReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "env": {
                    "description": "env,格式同创建app",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "app名称",
                    "type": "string"
                }
            }
        },
...
        "model.AppListBase": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AppInfo"
                    }
                }
            }
        },
        "model.AppListReq": {
            "type": "object",
            "required": [
                "current",
                "pageSize"
            ],
            "properties": {
                "category": {
                    "description": "类别",
                    "type": "string"
                },
                "current": {
                    "description": "当前第几页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页多少条",
                    "type": "integer"
                },
                "show_name": {
                    "description": "名称",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.AppListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.AppListBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.AppOpenReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "is_open": {
                    "description": "开启时传 true，关闭时传 false",
                    "type": "boolean"
                },
                "name": {
                    "description": "app名称",
                    "type": "string"
                }
            }
        },
        "model.AppReplicasReq": {
            "type": "object",
            "required": [
                "name",
                "replicas"
            ],
            "properties": {
                "name": {
                    "description": "app名称，注意不是show_name",
                    "type": "string"
                },
                "replicas": {
                    "type": "integer"
                }
            }
        },
        "model.AppRestartReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "app名称，注意不是show_name",
                    "type": "string"
                }
            }
        },
        "model.AppRestoreReq": {
            "type": "object",
            "required": [
                "app_name",
                "snapshot_id"
            ],
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "snapshot_id": {
                    "description": "快照id",
                    "type": "integer"
                }
            }
        },
        "model.AppScaleReq": {
            "type": "object",
            "required": [
                "cpu",
                "mem",
                "name"
            ],
            "properties": {
                "cpu": {
                    "description": "cpu核心数，例如1，2，3，4",
                    "type": "string"
                },
                "mem": {
                    "description": "内存大小，例如1，2，3，4",
                    "type": "string"
                },
                "name": {
                    "description": "app名称",
                    "type": "string"
                }
            }
        },
...
        "model.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.SC": {
            "type": "object",
            "properties": {
                "allow_volume_expansion": {
                    "description": "是否允许扩容",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.SCListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SC"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}